requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
groq==0.4.1
python-docx==1.1.0
reportlab==4.0.7
Pillow==10.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
Werkzeug==3.0.1